move transduce into transducer module, use steel derive, remove panic
also add spans for the errors. also creates a previous_span fn on VmCore
for easier spans and error handling.
